cosmetic edits for trace manager

+++ src/org/executequery/gui/browser/managment/tracemanager/TablePanel.java
@@ -232,7 +232,7 @@ public class TablePanel extends JPanel {
         logListPanel.setVerticalScrollBarPolicy(ScrollPaneConstants.VERTICAL_SCROLLBAR_ALWAYS);
         table.setRowSorter(new CustomTableRowSorter(dataModel));
         table.setDefaultRenderer(Object.class, new CustomTableCellRenderer());
-        table.setDefaultRenderer(Integer.class, new CustomTableCellRenderer());
+        table.setDefaultRenderer(Long.class, new CustomTableCellRenderer());
         table.setDefaultRenderer(Timestamp.class, new StatementTimestampTableCellRenderer());
         logListPanel.setViewportView(table);
         table.setAutoResizeMode(JTable.AUTO_RESIZE_OFF);
